change imports for auth.ts move
