--- src/pages/api/time.ts
@@ -1,8 +1,0 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
-
-export default async function handler(
-  _req: NextApiRequest,
-  res: NextApiResponse
-) {
-  res.status(200).json({ time: new Date() })
-}
